Audit fixes: widen achievements bitmask + compact Stats tab
Post-Round-3 sweep found two issues. Both fixed here.

1. Achievement bitmask saturation (latent UB).
   kAchievementCount reached 32 -- the limit of the uint32_t
   `achievements_` field. The static_assert was `<= 32` and still
   passed, but adding any 33rd achievement would compute
   `1u << 32` (undefined behavior) and silently fail to set or
   check the bit. Widen everything to uint64_t:
     - is_unlocked / achievement_bit / unlock now take/return u64
     - Game::achievements() returns u64
     - achievements_ field becomes u64
     - static_assert tightened to `<= 64`
   Save format bump v15 -> v16 adds `uint32_t achievements_hi` at
   the END of SaveData so the v2 `achievements` field offset stays
   put. init() reassembles `((u64)hi << 32) | lo`; force_save()
   splits back. v15 saves migrate cleanly with the high word
   zero-initialised.

2. Stats tab vertical overflow (cosmetic bug).
   Round 3 added 4 conditional rows at the top of draw_menu_stats
   (Quest / Today / Best friend / Stories). With all of them
   present alongside the existing 9 info rows + sparkline + diary,
   the rendered text exited the play area's bottom edge by ~14 px
   and clipped the "Yesterday: ..." line. Drop line spacing for the
   upper info block from 12 to 10 px (`constexpr int kInfoStep = 10`)
   to claw back ~18 px -- enough to keep every row visible in the
   worst case.

New bitmask_smoke covers: bit values at ids 0/31/32/33/63, unlock
+ is_unlocked roundtrip, Game save/load preserves both words,
v15->v16 migrator zeros the high word. All 21 prior smokes still
pass after the widening.

+++ lib/tama_core/include/tama/game.h
@@ -174,7 +174,7 @@ class Game {
   const Pet& pet() const { return pet_; }
   const Settings& settings() const { return settings_; }
   Settings& mut_settings() { dirty_ = true; return settings_; }
-  uint32_t  achievements() const { return achievements_; }
+  uint64_t  achievements() const { return achievements_; }
   uint16_t  streak_days()  const { return streak_days_; }
   Weather   weather()      const { return (Weather)weather_; }
   Personality personality() const { return (Personality)personality_trait_; }
@@ -379,7 +379,7 @@ class Game {
 
   Pet      pet_;
   Settings settings_;
-  uint32_t achievements_  = 0;
+  uint64_t achievements_  = 0;
   uint16_t streak_days_   = 0;
   uint64_t streak_last_visit_unix_ms_ = 0;
   uint64_t last_save_real_unix_ms_    = 0;
